Advanced Projects/research ai agent/requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
duckduckgo-search
